feat: move homepage FAQ schema into (home) route group layout
